Stiched contig example
This example shows how a seq can be extended to satisfy the biogo.step
Equaler interface and so be used as a step. This enables a collection of
sequences to be used as a single linear sequence-like type. Note that
the Contig type does not satisfy the seq.Sequence interface, although it
is possible to do so by adding, Clone, New, Slice and SetSlice, although
it is not clear what Slice and SetSlice would mean in this context.